Move CreateUserName/DemographicsSchema to common/user.ts to break import cycle
The previous commit exported CreateUserName/DemographicsSchema from
users/schema.ts so families/schema.ts could reuse them. That introduced
a circular import: users/schema.ts already imports UserFamilyRoleSchema
from families/schema.ts, so adding the reverse direction caused
families/schema.ts to evaluate CreateUserDemographicsSchema as
`undefined` at module init time. Symptom: TypeError 'Cannot read
properties of undefined (reading optional)' in 23 unrelated test files
when their transitive imports touched the api-contract.

Fix: move both shared schemas to common/user.ts where every contract
already imports from. No contract imports from another contract, so no
cycle. The schemas live alongside their inputs (IDENTIFIER_WITH_SPACES,
FreeReducedLunchStatusSchema) which are already in common/. Both
users/schema.ts and families/schema.ts now import them from
common/user.ts directly.
