+++ packages/api-contract/src/v1/common/user.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod';
 import { PaginationQuerySchema, createPaginatedResponseSchema, createSortQuerySchema } from './query';
+import { IDENTIFIER_WITH_SPACES } from './regex';
 
 export const UserRoleSchema = z.enum([
   'administrator',
@@ -134,6 +135,35 @@ export const ClassTypeSchema = z.enum(['homeroom', 'scheduled', 'other']);
  * Schema for free/reduced lunch status
  */
 export const FreeReducedLunchStatusSchema = z.enum(['Free', 'Reduced', 'Paid']);
+
+/**
+ * Name shape shared across user-creation endpoints (POST /v1/users,
+ * POST /v1/families, POST /v1/families/:familyId/users).
+ *
+ * Lives in `common/` so any sibling contract can reuse it without introducing
+ * a cross-contract import cycle.
+ */
+export const CreateUserNameSchema = z.object({
+  first: z.string().regex(IDENTIFIER_WITH_SPACES),
+  middle: z.string().regex(IDENTIFIER_WITH_SPACES).optional(),
+  last: z.string().regex(IDENTIFIER_WITH_SPACES),
+});
+
+/**
+ * Optional demographic fields shared across user-creation endpoints.
+ *
+ * Lives in `common/` so any sibling contract can reuse it without introducing
+ * a cross-contract import cycle.
+ */
+export const CreateUserDemographicsSchema = z.object({
+  gender: z.string().nullable().optional(),
+  race: z.string().nullable().optional(),
+  statusEll: z.string().nullable().optional(),
+  statusFrl: FreeReducedLunchStatusSchema.nullable().optional(),
+  statusIep: z.string().nullable().optional(),
+  hispanicEthnicity: z.boolean().nullable().optional(),
+  homeLanguage: z.string().nullable().optional(),
+});
 
 // Export types for individual schemas
 export type AuthProvider = z.infer<typeof AuthProviderSchema>;
