+++ packages/api-contract/src/v1/users/schema.ts
@@ -6,9 +6,10 @@ import {
   SchoolLevelSchema,
   AuthProviderSchema,
   FreeReducedLunchStatusSchema,
+  CreateUserNameSchema,
+  CreateUserDemographicsSchema,
 } from '../common/user';
 import { UserFamilyRoleSchema } from '../families/schema';
-import { IDENTIFIER_WITH_SPACES } from '../common/regex';
 
 /**
  * Response schema for GET /users/:userId
@@ -80,33 +81,6 @@ const FamilyMembershipSchema = z.object({
 
 export const UserMembershipSchema = z.union([OrgMembershipSchema, FamilyMembershipSchema]);
 export type UserMembership = z.infer<typeof UserMembershipSchema>;
-
-/**
- * Name shape shared across user-creation endpoints (POST /v1/users,
- * POST /v1/families, POST /v1/families/:familyId/users).
- *
- * Exported so sibling contracts can reuse rather than duplicate.
- */
-export const CreateUserNameSchema = z.object({
-  first: z.string().regex(IDENTIFIER_WITH_SPACES),
-  middle: z.string().regex(IDENTIFIER_WITH_SPACES).optional(),
-  last: z.string().regex(IDENTIFIER_WITH_SPACES),
-});
-
-/**
- * Optional demographic fields shared across user-creation endpoints.
- *
- * Exported so sibling contracts can reuse rather than duplicate.
- */
-export const CreateUserDemographicsSchema = z.object({
-  gender: z.string().nullable().optional(),
-  race: z.string().nullable().optional(),
-  statusEll: z.string().nullable().optional(),
-  statusFrl: FreeReducedLunchStatusSchema.nullable().optional(),
-  statusIep: z.string().nullable().optional(),
-  hispanicEthnicity: z.boolean().nullable().optional(),
-  homeLanguage: z.string().nullable().optional(),
-});
 
 const CreateUserIdentifiersSchema = z.object({
   stateId: z.string().optional(),
